Fix getchildren()
Nodes in memory can be modified in cache but not in the DB within the
same iteration of the events loop. It happens that parent-child relation
has changed but not reflected on DB, but the previous `getchildren()`
looks for the handles of child nodes in DB, getting outdated info.

The workaround is to move `getchildren()` to the cache of nodes, where
the cache can check if there is any node not being child node anymore,
or the other way, a node which wasn't a child and now it is.