First draft at refactoring Transfer states
Motivated by discovering a bug where the TransferSlot was deleted, but state was still TRANSFERSTATE_ACTIVE - which causes it to stall forever.  To reproduce, send `t:[]` with putfile, and convert the error in the reply from -2 to -17.

State field and Slot to be private and managed only by the transfer in new function setState

Made file opening (sync or async) the TranferSlot's responsibility to simplify the combination of states - before we could have a slot, which could have an fa.  Asyncopencontext is now in the slot rather than the transfer too which is more natural.  Logic has been simplified in a few places as a result

Documented the transfer states

Discovered a secret unnamed state when file attributes are not attached yet when the file content uploading part has already completed.

It seems that the RETRYING state sometimes had a slot still and sometimes did not, sometimes had retrying=true, sometimes not - that still needs sorting out (coded for always no slot now)